Fix all gyroid (haniwa) crashes and animation bugs on 64-bit
This patch extends the initial 64-bit audio fixes to resolve remaining
crashes with specific gyroid types:

1. Asset Header Endianness: Fixed Big-Endian ROM header parsing for
   ArcHeader and ArcEntry in Nas_BankHeaderInit. On PC, these are now
   byte-swapped in-place so asset addresses and sizes are read correctly.
2. Filter Endianness: Subtrack filter taps (8 coefficients) in sequence
   data are now byte-swapped to host order in track.c, preventing audio
   corruption and crashes during interaction.
3. 64-bit Command Safety: Updated AUDIOCMD_SET_EXTERNAL_POINTER to use
   full-width asVoidPtr to prevent 64-bit pointer truncation when passing
   data to the audio engine.
4. Robustness: Added safety checks to Na_SysLevStop and fixed an infinite
   loop/OOB access in Na_SysLevStart.
5. Build Fix: Included pc_bswap.h in track.c for missing definitions.

These changes combined with the previous pointer-size modernization ensure
that every gyroid type functions correctly without crashing the game.

+++ include/jaudio_NES/audiothread.h
@@ -19,7 +19,7 @@ extern "C" {
 
 extern void DspSyncCountClear(int count);
 extern int DspSyncCountCheck(void);
-extern void StartAudioThread(void* pHeap, size_t heapSize, u32 aramSize, u32 flags);
+extern void StartAudioThread(void* pHeap, s32 heapSize, u32 aramSize, u32 flags);
 
 #ifdef TARGET_PC
 void pc_audio_process_frame(void);
