+++ include/jaudio_NES/verysimple.h
@@ -6,7 +6,7 @@
 extern "C" {
 #endif
 
-void Jac_Start(void*, size_t, u32);
+void Jac_Start(void*, u32, u32);
 void Jac_PlayInit(void);
 
 #ifdef __cplusplus
